+++ catalog/catalog-app/pom.xml
@@ -671,17 +671,17 @@
             <classifier>features</classifier>
         </dependency>
         <dependency>
-            <groupId>ddf.migration</groupId>
+            <groupId>ddf.catalog.migrate</groupId>
             <artifactId>migration-api</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>ddf.migration</groupId>
-            <artifactId>migration-commands</artifactId>
+            <groupId>ddf.catalog.migrate</groupId>
+            <artifactId>migration-command</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>ddf.migration</groupId>
+            <groupId>ddf.catalog.migrate</groupId>
             <artifactId>workspace-query-separation</artifactId>
             <version>${project.version}</version>
         </dependency>
