+++ catalog/migrate/workspace-query-separation/src/test/resources/query.xml
