+++ catalog/pom.xml
@@ -524,17 +524,17 @@
                 <version>${project.version}</version>
             </dependency>
             <dependency>
-                <groupId>ddf.migration</groupId>
+                <groupId>ddf.catalog.migrate</groupId>
                 <artifactId>migration-api</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
-                <groupId>ddf.migration</groupId>
+                <groupId>ddf.catalog.migrate</groupId>
                 <artifactId>migration-commands</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
-                <groupId>ddf.migration</groupId>
+                <groupId>ddf.catalog.migrate</groupId>
                 <artifactId>workspace-query-separation</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -560,6 +560,7 @@
         <module>validator</module>
         <module>confluence</module>
         <module>async</module>
+        <module>migrate</module>
     </modules>
     <build>
         <plugins>
